Rename index router -> home
