+++ controllers/home.js
@@ -1,7 +1,7 @@
-var HomePageController = {
+var HomeController = {
   Index: function(req, res) {
     res.render('index', { title: 'Acebook' });
   }
 };
 
-module.exports = HomePageController;
+module.exports = HomeController;
