+++ routes/home.js
@@ -1,0 +1,8 @@
+var express = require('express');
+var router = express.Router();
+
+var HomeController = require('../controllers/home');
+
+router.get('/', HomeController.Index);
+
+module.exports = router;
