+++ routes/posts.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 
-var PostsController = require('../controllers/posts_controller')
+var PostsController = require('../controllers/posts')
 
 router.get('/', PostsController.Index);
 router.post('/', PostsController.Create);
